Removed NotificationList, MyPullRequests moved to Home page
