--- src/client/components/NotificationList/NotificationList.stories.js
@@ -1,12 +1,0 @@
-/* @flow */
-import React from 'react'
-import { storiesOf } from '@kadira/storybook'
-import { muiTheme } from 'storybook-addon-material-ui'
-
-import NotificationList from './NotificationList.js'
-
-storiesOf('NotificationList', module)
-  .addDecorator(muiTheme())
-  .add('default', () => (
-    <NotificationList />
-  ))
