--- src/client/components/NotificationList/index.js
@@ -1,3 +1,0 @@
-/* @flow */
-import NotificationList from './NotificationList'
-export default NotificationList
